recommender/requirements.txt
pandas
pyarrow          # for reading Parquet
openai
kafka-python    # for Kafka producer/consumer